Update mk_links to backup existing files
mv existing files to .bak and create symlinks to all dotfiles
Previously, this file moved the existing dotfiles to the .dotfiles
folder and then creating symlinks.  This was useful only in creating
the .dotfiles folder.  Now it can be run after the .dotfiles folder is cloned

Changes to be committed:
      modified:   mk_links